+++ cdap-data-fabric/src/main/java/io/cdap/cdap/store/RepositoryStore.java
@@ -16,37 +16,83 @@
 
 package io.cdap.cdap.store;
 
-import com.google.inject.ImplementedBy;
 import io.cdap.cdap.common.NamespaceNotFoundException;
 import io.cdap.cdap.common.RepositoryNotFoundException;
 import io.cdap.cdap.proto.id.NamespaceId;
 import io.cdap.cdap.proto.sourcecontrol.RepositoryConfig;
 import io.cdap.cdap.proto.sourcecontrol.RepositoryMeta;
+import io.cdap.cdap.spi.data.StructuredTableContext;
+import io.cdap.cdap.spi.data.TableNotFoundException;
+import io.cdap.cdap.spi.data.transaction.TransactionRunner;
+import io.cdap.cdap.spi.data.transaction.TransactionRunners;
+
+import javax.inject.Inject;
 
 /**
- * Store for repository configuration. This interface is by default implemented by {@link DefaultRepositoryStore},
- * unless explicitly overridden for testing.
+ * Store for repository configuration.
  */
-@ImplementedBy(DefaultRepositoryStore.class)
-public interface RepositoryStore {
+public class RepositoryStore {
+
+  private final TransactionRunner transactionRunner;
+
+  @Inject
+  RepositoryStore(TransactionRunner transactionRunner) {
+    this.transactionRunner = transactionRunner;
+  }
+
+  private RepositoryTable getRepositoryTable(StructuredTableContext context) throws TableNotFoundException {
+    return new RepositoryTable(context);
+  }
+
+  private NamespaceTable getNamespaceTable(StructuredTableContext context) throws TableNotFoundException {
+    return new NamespaceTable(context);
+  }
+
   /**
    * Sets the repository configuration foe the namespace.
    * @param namespace {@link NamespaceId} The namespace where the repository belongs
    * @param repository The {@link RepositoryConfig} that is to be set in table
    * @return {@link RepositoryMeta}
    */
-  RepositoryMeta setRepository(NamespaceId namespace, RepositoryConfig repository) throws NamespaceNotFoundException;
+  public RepositoryMeta setRepository(NamespaceId namespace, RepositoryConfig repository)
+    throws NamespaceNotFoundException {
+    return TransactionRunners.run(transactionRunner, context -> {
+      NamespaceTable nsTable = getNamespaceTable(context);
+      if (nsTable.get(namespace) == null) {
+        throw new NamespaceNotFoundException(namespace);
+      }
+
+      RepositoryTable repoTable = getRepositoryTable(context);
+      repoTable.create(namespace, repository);
+      return repoTable.get(namespace);
+    }, NamespaceNotFoundException.class);
+  }
 
   /**
    * Deletes the repository configuration of the namespace.
    * @param namespace {@link NamespaceId} The namespace of the repository to delete
    */
-  void deleteRepository(NamespaceId namespace);
+  public void deleteRepository(NamespaceId namespace) {
+    TransactionRunners.run(transactionRunner, context -> {
+      RepositoryTable repoTable = getRepositoryTable(context);
+      repoTable.delete(namespace);
+    });
+  }
 
   /**
-   * 
+   *
    * @param namespace {@link NamespaceId} The namespace of the repository to get
    * @return {@link RepositoryMeta}
    */
-  RepositoryMeta getRepositoryMeta(NamespaceId namespace) throws RepositoryNotFoundException;
+  public RepositoryMeta getRepositoryMeta(NamespaceId namespace) throws RepositoryNotFoundException {
+    return TransactionRunners.run(transactionRunner, context -> {
+      RepositoryTable table = getRepositoryTable(context);
+      RepositoryMeta repoMeta = table.get(namespace);
+      if (repoMeta == null) {
+        throw new RepositoryNotFoundException(namespace);
+      }
+
+      return repoMeta;
+    }, RepositoryNotFoundException.class);
+  }
 }
